Move our libraries into a "Spectrogram" directory
