+++ src/CMakeLists.txt
@@ -1,5 +1,2 @@
 
-add_subdirectory(Audio)
-add_subdirectory(Fourier)
-add_subdirectory(Graphics)
-
+add_subdirectory(Spectrogram)
