+++ src/Spectrogram/CMakeLists.txt
@@ -1,0 +1,5 @@
+
+add_subdirectory(Audio)
+add_subdirectory(Fourier)
+add_subdirectory(Graphics)
+
