+++ src/Spectrogram/Graphics/placeholder.cpp
