+++ src/Spectrogram/Graphics/placeholder.h
